requirements.txt
Django==2.1.5
django-articles==2.4.1
django-classy-tags==0.8.0
django-cms==3.6.0
django-filer==1.4.4
django-formtools==2.1
django-js-asset==1.1.0
django-mptt==0.9.1
django-polymorphic==2.0.3
django-sekizai==0.10.0
django-treebeard==4.3
<<<<<<< HEAD
=======
djangocms-admin-style==1.3.0
djangocms-attributes-field==1.0.0
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-googlemap==1.2.0
djangocms-link==2.3.1
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
>>>>>>> 2182f3a4b8fdddf73aa0b260257e2d329ad6c815
easy-thumbnails==2.5
html5lib==1.0.1
mysqlclient==1.4.1
Pillow==5.4.1
pytz==2018.9
six==1.12.0
Unidecode==1.0.23
webencodings==0.5.1
